+++ src/index.js
@@ -2,9 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import './index.css';
-import App from './App';
 import * as serviceWorker from './serviceWorker';
 import store from './store';
+import App from './components/App';
 
 ReactDOM.render(
   <Provider store={store}>
@@ -12,6 +12,7 @@ ReactDOM.render(
   </Provider>,
   document.getElementById('root')
 );
+
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
